requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-SocketIO==5.3.6
SQLAlchemy==2.0.30
Werkzeug==3.0.3
python-dotenv==1.0.1
qrcode[pil]==7.4.2
PyJWT==2.8.0
openpyxl==3.1.3
gunicorn==22.0.0
eventlet==0.36.1
psycopg2-binary==2.9.9
APScheduler==3.10.4
anthropic==0.25.0
